refactor(naming): rename `services` folder to `data-access`
