+++ src/app/features/settings/models/settings-vm.model.ts
@@ -1,5 +1,5 @@
 import { IUserSettingsState } from '@fts-store/user-settings';
-import { ISettingsFeatureState } from '../settings.store';
+import { ISettingsFeatureState } from '../data-access/settings.store';
 
 export interface ISettingsFeatureVM {
   userSettings: IUserSettingsState;
